% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{Bootstrap Interval for a Two Sample Test}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{x1}{#First sample}

\item{x2}{#Second sample}

\item{conf.level}{#Confidence Level; alpha}

\item{iter}{#the number of iterations}

\item{mudiff}{#The difference in sample means}

\item{test}{#The type of test to be executed}
}
\value{
#The p-value assocaiated with the test
}
\description{
Bootstrap Interval for a Two Sample Test
}
\examples{
set.seed(30);x=rnorm(15,mean=10,sd=7)
set.seed(40);y=rnorm(20,mean=12,sd=4)

boot2pval(x1=x,x2=y)
}
