% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyOutliers.R
\name{identityOutliers}
\alias{identityOutliers}
\title{Identify Outliers Function}
\usage{
identityOutliers(vector)
}
\arguments{
\item{vector}{}
}
\value{
a list of length two. The first element of the list comtains all values that are possibly outliers, while the second element contain all values that are definitely outliers
}
\description{
Identify Outliers Function
}
\examples{
x <- c(4,90,100,95); identifyOutliers(x)
}
