% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Experiment Simulation}
\usage{
mybin(iter, n, p)
}
\arguments{
\item{p}{}
}
\value{
The Bionomial Experiment plot, and n values
}
\description{
Binomial Experiment Simulation
}
\examples{
iter = 100
n = 5
p = 0.7
mybin(100,5,0.7)

}
