% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample size}

\item{fun}{type of function}

\item{alpha}{significance level, probability of rejecting the null hypothesis when it is true.}

\item{cx}{number indicating the amount by which plotting text and symbols should be scaled relative to the default}

\item{...}{}
}
\value{

}
\description{
This function creates samples of size n with replacement using the bootstrap procedure.
}
\examples{
For a bootstrap interval of the mean with reproducible random numbers set.seed(68),
and a sample that takes a normal distribution with mean 10 and standard deviation 4, sam=rnorm(20,mean=10,sd=4).
To create a histogram of bootstrap samples of the function mean with 10000 iterations of 95\% CI,
call: myboot2(10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="cadetblue3",cx=1.5)
}
