% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{95% Confidence Interval of the Mean}
\usage{
myci(x)
}
\arguments{
\item{x}{a single set of sample statistics}
}
\value{
a vector of length 2; the first value is the lower limit and the second its upper limit
}
\description{
95% Confidence Interval of the Mean
}
\examples{
myci(c(6,7,5,4,5,6,7,5,4,4,5,5,6,5,4,4,4,5))
}
