% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Binomial Distribution (Central Limit Theorm)}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{p}{probability}

\item{...}{}
}
\value{
a histogram showcasing the binomial distributions' sample means
}
\description{
Binomial Distribution (Central Limit Theorm)
}
\examples{
myclbt(n = 20, iter = 10000)
}
