% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations, or times the graph will run}

\item{lambda}{poisson parameter, = total number of events (k) / number of units (n)}

\item{...}{}
}
\value{

}
\description{
A function for a poisson distribution that produces a  histogram of sample mean based on sample size, iteration and lambda; barplot of the sampled y value; and probability function for poisson.
}
\examples{
for n = 40, iter=40, and lambda = 4, call: mycltp(n=40,iter=40,lambda=4)
}
