% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myF.R
\name{myF}
\alias{myF}
\title{Square Function}
\usage{
myF(x)
}
\arguments{
\item{x}{}
}
\value{
a vector of squared components in the form of a vector
}
\description{
Takes a vector and returns a vector of squared components.
}
