% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Normal Distribution Maximum Likelihood Estimates}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{#data}

\item{mu}{#a sequence of possible means}

\item{sig}{#a sequence of possible standard deviations}

\item{...}{#Additional parameters}
}
\value{
#a list containing x, coordinates of estimated mu & sigma, and max likelihood
}
\description{
Makes maximum likelihood estimations about a normal distributions mean and standard deviation.
}
\examples{
mymlnorm(x = c(4,5,6,7,4), mu = seq(1, 4, 10000), sig = seq(.1, 3, 10000), ...)
}
