% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{population mean}

\item{sigma}{population variance}

\item{a}{upper limit of the curve}
}
\value{

}
\description{
A function that produces  a curve with the area shaded between the curve and the x axis from - infinity to x=a and will calculte the area P(X<=a)
}
\examples{
If my=10, sigma=5, and a=6, then invoke:: myncurve(mu=10,sigma=5,a=6)
}
