% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysample Function}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Sample size}

\item{iter}{number of iterations}

\item{time}{amount of time between plots being displayed}
}
\value{
a
}
\description{
This function will make a plot for each random sample generated; one for each iteration is specified
}
