% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Poincurvx3.R
\name{Poincurvx3}
\alias{Poincurvx3}
\title{CLT Poisson}
\usage{
Poincurvx3(n, iter, lambda = 10, ...)
}
\arguments{
\item{...}{}
}
\value{
3 plots: A Histogram with a normal curve of the sample means and the density of the random values created by the function
A histogram of the values and their relative frequency and A histogram of the values and their probabilities.
}
\description{
CLT will work with discrete or continuous distributions my Central Limit Function. Creates a random
Vector of values, enters them into a matriz and then creates 3 histograms to evaluate how normal the functions are
}
\examples{
n = 10
iter = 100000
lambda = 4
Poincurvx3(10,10000,4)


}
