% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simbin.R
\name{simbin}
\alias{simbin}
\title{Binomial Experiment Simulation}
\usage{
simbin(iter, n, p)
}
\arguments{
\item{p}{}
}
\value{
The Bionomial Experiment plot, and n values
}
\description{
The Binomial Experiment Simulation function comes from Lab 5. This function produces
a simulation of the Binomial experiment by placing random values of size n, iterated
iter times with a probability of p.
}
\examples{
iter = 100
n = 5
p = 0.7
simbin(100,5,0.7)

}
