% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS.R
\name{TSS}
\alias{TSS}
\title{Total Sum of Squares}
\usage{
TSS(x)
}
\arguments{
\item{x}{}
}
\value{
The total sum of squares value
}
\description{
The total sum of squares function takes a variable of read in data and gives the sum of the square deviation
scores from the mean variable. It is the error calculation of the no predictor model.
}
\examples{
x = (1:20)
TSS(x)
}
