% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot2pval.R
\name{boot2pval}
\alias{boot2pval}
\title{Bootstrap testing for two samples}
\usage{
boot2pval(x1, x2, conf.level = 0.95, iter = 3000, mudiff = 0, test = "two")
}
\arguments{
\item{test}{}
}
\value{
histogram of tstats of samples and the p-value for the difference between the means
}
\description{
This function takes two samples and returns the p-value of the two samples,
along with a confidence interval of the difference between the means of the samples.
Then the function returns a histogram with the calculated t-statistics.
}
\examples{

x1 = c(2,3,4,5)
x2 = c(4,5,8,9)
mudiff = 2
boot2pval(x1,x2,mudiff)

}
