% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample object}

\item{fun}{function to apply to sample}

\item{alpha}{alpha (used to manipulate the level of confidence)}

\item{cx}{final character size}

\item{...}{further parameters}
}
\value{
a histogram showcasing the inputted distributions' sample statistic as well as bootsrapped confidence interval
}
\description{
Bootstrap Confidence Interval
}
\examples{
myboot2(x = rnorm(20,mean=10,sd=4), alpha = 0.2)
}
