% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{...}{}
}
\value{
returns a historgram with an interval of the range of means within a percentage certainity
}
\description{
This function takes a variety of parameters and creates a histogram from a random sample (x) and finds the
interval for the range of means within a percentage of certainty.
}
\examples{
iter = 10000
x = c(1,3,4,5,6,7,2,1,3,7)
fun = "mean"
alpha = 0.05
cx = 1.5
myboot2(iter,x,fun,alpha,cx)

}
