% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval Function for One Sample}
\usage{
myci(x, conf = 0.95, ...)
}
\arguments{
\item{...}{}
}
\value{
95% confidence interval, ci of the sample
}
\description{
Takes a sample of values and returns a 95% confidence (default) interval of the mean of the sample.
}
\examples{
x = rnorm(30,mean = 10,sd=12)
myci(x)

}
