% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve Function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{a}{}
}
\value{
a normal curve, a shaded normal curve within the area of the probability, and the area
}
\description{
Creates a normal curve, the curve with the area shaded based upon a probability using ggplot
and the area.
}
\examples{
mu = 10
sigma = 5
a = 6
myncurve(10,5,6)


}
