% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simbin.R
\name{simbin}
\alias{simbin}
\title{Binomial Experiment Simulation}
\usage{
simbin(iter, n, p)
}
\arguments{
\item{p}{}
}
\value{
The Bionomial Experiment plot, and n values
}
\description{
Binomial Experiment Simulation
}
\examples{
iter = 100
n = 5
p = 0.7
simbin(100,5,0.7)

}
